requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
djangorestframework==3.9.1
pytz==2018.9
